[pkg/stanza] Keep files open between polls on Windows
